+++ Access/src/utopia/access/model/event/ServerSentEvent.scala
@@ -1,6 +1,5 @@
 package utopia.access.model.event
 
-import utopia.flow.collection.immutable.Empty
 import utopia.flow.generic.model.immutable.Value
 import utopia.flow.operator.MaybeEmpty
 
@@ -9,11 +8,15 @@ import utopia.flow.operator.MaybeEmpty
   * @author Mikko Hilpinen
   * @since 30.03.2025, v1.6
   * @param eventType Type of this event. May be empty if this event doesn't have a type.
-  * @param data Contents of this event. Each line is represented with a separate entry.
+  * @param data Contents of this event.
+  *             If this event consists of multiple data entries, these are represented with a
+  *             Value of type VectorType or PairType.
+  * @param dataEntryCount Number of data entries represented with 'data'
   * @param id Id of this event. May be empty.
   * @param completed Whether this represents a completed event (true) or a snapshot (false)
   */
-case class ServerSentEvent(eventType: String = "", data: Seq[Value] = Empty, id: String = "", completed: Boolean = true)
+case class ServerSentEvent(eventType: String = "", data: Value = Value.empty, dataEntryCount: Int = 0,
+                           id: String = "", completed: Boolean = true)
 	extends MaybeEmpty[ServerSentEvent]
 {
 	// COMPUTED -------------------------